+++ backend/src/main/java/com/aliados/backend/dto/MudanzaTierResponseDTO.java
@@ -2,15 +2,17 @@ package com.aliados.backend.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class MudanzaTierResponseDTO {
 
     private Long id;
     private String nombre;
     private String emoji;
-    private Double precioBase;
+    private BigDecimal precioBase;
     private Integer minutosIncluidos;
-    private Double precioBloque30Min;
+    private BigDecimal precioBloque30Min;
     private String descripcion;
     private String descripcionCompleta;
     private Integer orden;
